Simplify and clean up the installation scripts (#21517)
* Rename backend test coverage script and reduce unnecessary logging.

* Drop unused run_tests script

* Move orphaned scriptsstatusimport_test into install_third_party_libs_test

* Drop change detector test

* Get rid of setup.py, setup_gae.py, install_third_party.py

* Fix tests.

* Fix lint issues.

* Fix import error

* Fix lint issues

* Fix mypy issues

* Tidy output

* Fix more type errors

* Fix more type errors

* Fix tests

* Fix tests

* Continue fixing tests

* Add --strip-extras flag

* Fix test in servers_test opening a browser during the tests. Remove common.py dep from install_dependencies_json_packages since it blocks the docker installation.

* Add dir creation for dependencies in third_party

* fix imports

* fix lint

* fix backend test

* Fix coverage

* Fix coverage
