+++ .github/workflows/backend_unit_tests.yml
@@ -135,7 +135,7 @@ jobs:
       - name: Check coverage
         if: startsWith(github.head_ref, 'update-changelog-for-release') == false
         shell: bash
-        run: make run_tests.check_overall_backend_test_coverage
+        run: make run_tests.check_backend_test_coverage
       - name: Report failure if failed on oppia/oppia develop branch
         if: ${{ failure() && github.event_name == 'push' && github.repository == 'oppia/oppia' && github.ref == 'refs/heads/develop'}}
         uses: ./.github/actions/send-webhook-notification
