+++ Makefile
@@ -155,9 +155,9 @@ run_tests.backend: ## Runs the backend tests
 	@echo '------------------------------------------------------'
 	$(MAKE) stop
 
-run_tests.check_overall_backend_test_coverage: ## Runs the check for overall backend test coverage
+run_tests.check_backend_test_coverage: ## Runs the check for overall backend test coverage
 	$(MAKE) start-devserver
-	$(SHELL_PREFIX) dev-server python -m scripts.check_overall_backend_test_coverage
+	$(SHELL_PREFIX) dev-server python -m scripts.check_backend_test_coverage
 	$(MAKE) stop
 
 run_tests.frontend: ## Runs the frontend unit tests
