+++ requirements.txt
@@ -6,9 +6,9 @@
 # This file is autogenerated by pip-compile with Python 3.9
 # by the following command:
 #
-#    pip-compile --generate-hashes --no-emit-index-url --output-file=requirements.txt requirements.in
+#    pip-compile --generate-hashes --no-emit-index-url --output-file=requirements.txt --strip-extras requirements.in
 #
-apache-beam[gcp]==2.43.0 \
+apache-beam==2.43.0 \
     --hash=sha256:046ef2fe4dde1a4e17709fc6dd38f4401e3f83218388ddc491ee910020802c89 \
     --hash=sha256:07ba2bf2832614a221ca8d0bc144d0e2e15e68d3ccd8204c8990e0ae2b8a452e \
     --hash=sha256:07fe0d1589eea431e271d3fa0a25f059f003afd76d45b0713c15954e88ca2208 \
@@ -231,7 +231,7 @@ firebase-admin==5.2.0 \
 future==0.18.3 \
     --hash=sha256:34a17436ed1e96697a86f9de3d15a3b0be01d8bc8de9c1dffd59fb8234ed5307
     # via -r requirements.in
-google-api-core[grpc]==2.23.0 \
+google-api-core==2.23.0 \
     --hash=sha256:2ceb087315e6af43f256704b871d99326b1f12a9d6ce99beaedec99ba26a0ace \
     --hash=sha256:c20100d4c4c41070cf365f1d8ddf5365915291b5eb11b83829fbd1c999b5122f
     # via
@@ -476,7 +476,7 @@ google-resumable-media==2.7.2 \
     # via
     #   google-cloud-bigquery
     #   google-cloud-storage
-googleapis-common-protos[grpc]==1.63.2 \
+googleapis-common-protos==1.63.2 \
     --hash=sha256:27a2499c7e8aff199665b22741997e485eccc8645aa9176c7c988e6fae507945 \
     --hash=sha256:27c5abdffc4911f28101e635de1533fb4cfd2c37fbaa9174587c799fac90aa87
     # via
@@ -1523,5 +1523,6 @@ zstandard==0.23.0 \
     # via apache-beam
 
 # WARNING: The following packages were not pinned, but pip requires them to be
-# pinned when the requirements file includes hashes. Consider using the --allow-unsafe flag.
+# pinned when the requirements file includes hashes and the requirement is not
+# satisfied by a package already installed. Consider using the --allow-unsafe flag.
 # setuptools
