+++ requirements_dev.txt
@@ -2,7 +2,7 @@
 # This file is autogenerated by pip-compile with Python 3.9
 # by the following command:
 #
-#    pip-compile --generate-hashes --no-emit-index-url --output-file=requirements_dev.txt requirements_dev.in
+#    pip-compile --generate-hashes --no-emit-index-url --output-file=requirements_dev.txt --strip-extras requirements_dev.in
 #
 astroid==2.15.6 \
     --hash=sha256:389656ca57b6108f939cf5d2f9a2a825a3be50ba9d589670f393236e0a03b91c \
