+++ scripts/backend_test_shards.json
@@ -222,8 +222,6 @@
         "core.storage.suggestion.gae_models_test",
         "core.storage.auth.gae_models_test",
         "scripts.linters.html_linter_test",
-        "scripts.script_import_test",
-        "scripts.setup_gae_test",
         "scripts.linters.general_purpose_linter_test",
         "core.storage.app_feedback_report.gae_models_test",
         "core.jobs.batch_jobs.model_validation_jobs_test",
@@ -285,10 +283,8 @@
         "core.storage.recommendations.gae_models_test",
         "scripts.linters.other_files_linter_test",
         "scripts.install_dependencies_json_packages_test",
-        "scripts.install_third_party_test",
         "scripts.check_frontend_test_coverage_test",
         "core.platform.models_test",
-        "scripts.setup_test",
         "core.storage.feedback.gae_models_test",
         "scripts.pre_commit_hook_test",
         "core.storage.activity.gae_models_test",
@@ -300,7 +296,7 @@
         "core.storage.statistics.gae_models_test",
         "core.storage.exploration.gae_models_test",
         "scripts.release_scripts.repo_specific_changes_fetcher_test",
-        "scripts.docstrings_checker_test",
+        "scripts.linters.docstrings_checker_test",
         "extensions.domain_test",
         "core.controllers.learner_dashboard_test",
         "core.controllers.feedback_updates_test",
@@ -336,7 +332,7 @@
         "scripts.run_lighthouse_tests_test",
         "scripts.start_test",
         "setup_test",
-        "scripts.check_overall_backend_test_coverage_test",
+        "scripts.check_backend_test_coverage_test",
         "core.feconf_test",
         "core.domain.cron_services_test",
         "scripts.run_portserver_test",
@@ -354,7 +350,6 @@
         "scripts.check_backend_associated_test_file_test",
         "scripts.run_custom_eslint_tests_test",
         "scripts.run_presubmit_checks_test",
-        "scripts.run_tests_test",
         "core.controllers.collection_viewer_test",
         "core.domain.rating_services_test",
         "core.controllers.recent_commits_test",
