+++ scripts/pre_push_hook.py
@@ -275,7 +275,7 @@ def check_for_backend_python_library_inconsistencies() -> None:
         print('\n')
         common.print_each_string_after_two_new_lines([
             'Please fix these discrepancies by editing the `requirements.in`\n'
-            'file or running `scripts.install_third_party` to regenerate\n'
+            'file or running `scripts.install_third_party_libs` to regenerate\n'
             'the `third_party/python_libs` directory.\n'])
         sys.exit(1)
     else:
