+++ scripts/run_mypy_checks.py
@@ -184,14 +184,14 @@ def main(args: Optional[List[str]] = None) -> int:
     if not feconf.OPPIA_IS_DOCKERIZED:
         install_third_party_libraries(parsed_args.skip_install)
 
-        print('Installing Mypy and stubs for third party libraries.')
+        print('Installing Mypy and stubs for third party libraries...')
         return_code, mypy_exec_path = install_mypy_prerequisites(
             parsed_args.install_globally)
         if return_code != 0:
-            print('Cannot install Mypy and stubs for third party libraries.')
+            print('Installation failed.')
             sys.exit(1)
 
-        print('Installed Mypy and stubs for third party libraries.')
+        print('Installation succeeded.')
 
     mypy_exec_path = os.path.join(MYPY_TOOLS_DIR, 'bin', 'mypy')
